requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.6.3
Brotli==1.1.0
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
kaitaistruct==0.10
mysql-connector-python==9.3.0
numpy==2.2.6
outcome==1.3.0.post0
pandas==2.2.3
pyasn1==0.6.1
pycparser==2.22
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
selenium==4.33.0
selenium-wire==5.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
zstandard==0.23.0
